feat: serve default locale at root
